+++ src/Ext/House/Hooks.cpp
@@ -328,3 +328,38 @@ DEFINE_HOOK(0x4FD8F7, HouseClass_UpdateAI_OnLastLegs, 0x10)
 
 	return SkipGameCode;
 }
+
+DEFINE_HOOK(0x4FD77C, HouseClass_ExpertAI_Superweapons, 0x5)
+{
+	enum { SkipSWProcess = 0x4FD7A0 };
+
+	if (RulesExt::Global()->AISuperWeaponDelay.isset())
+		return SkipSWProcess;
+
+	return 0;
+}
+
+DEFINE_HOOK(0x4F9038, HouseClass_AI_Superweapons, 0x5)
+{
+	GET(HouseClass*, pThis, ESI);
+
+	if (!RulesExt::Global()->AISuperWeaponDelay.isset() || pThis->IsControlledByHuman() || pThis->Type->MultiplayPassive)
+		return 0;
+
+	int delay = RulesExt::Global()->AISuperWeaponDelay.Get();
+
+	if (delay > 0)
+	{
+		auto const pExt = HouseExt::ExtMap.Find(pThis);
+
+		if (pExt->AISuperWeaponDelayTimer.HasTimeLeft())
+			return 0;
+
+		pExt->AISuperWeaponDelayTimer.Start(delay);
+	}
+
+	if (!SessionClass::IsCampaign() || pThis->IQLevel2 >= RulesClass::Instance->SuperWeapons)
+		pThis->AI_TryFireSW();
+
+	return 0;
+}